Make the "choice" in weighted and quadratic voting as a string to double mapping. Fix crash when displaying String value of user vote.

+++ Goverland/Goverland/Snapshot Proposal/Vote Tab/Votes/SnapshotAllVotesView.swift
@@ -22,7 +22,7 @@ struct SnapshotAllVotesView: View {
             case .approval, .rankedChoice:
                 _SnapshotAllVotesView<[Int]>(proposal: proposal)
             case .weighted, .quadratic:
-                _SnapshotAllVotesView<[String: Int]>(proposal: proposal)
+                _SnapshotAllVotesView<[String: Double]>(proposal: proposal)
             }
         }
     }
